refactor: extract MarketMindAIPage data logic into useMarketMindAIData hook (F5)
Move the chat/workspace fetch, streaming, and state logic (with its
TEMPLATE_KEY and formatActionError helpers) into a page-local hook
(components/marketmind-ai/useMarketMindAIData.js). MarketMindAIPage becomes
a render shell (771 -> 368 lines) keeping only its JSX and presentation
helpers (class-name maps, getSentimentToneWord).

Behavior-preserving: hook body moved verbatim (same hook order incl.
useEffectEvent/useAuth, same effect deps), no early returns, returning
exactly the state and handlers the JSX consumes. Verified by the existing
5-test suite, full suite (83), lint, and build.
